refactored files to have openAPI specs
