+++ scheme-api/open-api-specification.yml
@@ -1,21 +1,28 @@
 openapi: "3.0.0"
 info:
   version: 1.0.0
-  title: Items API
+  title: Fee Scheme API
 paths:
-  /api/v1/items:
+  /api/v1/category-of-law/{feeCode}:
     get:
       tags:
-        - items
-      summary: Get items
-      operationId: getItems
+        - categoryOfLaw
+      parameters:
+        - name: feeCode
+          in: path
+          required: true
+          description: Fee Code to retrieve Category of Law
+          schema:
+            type: string
+      summary: Get Category of Law
+      operationId: getCategoryOfLaw
       responses:
         '200':
           description: 'Success'
           content:
             application/json:
               schema:
-                $ref: "#/components/schemas/Items"
+                $ref: "#/components/schemas/CategoryOfLaw"
         '400':
           description: 'Bad request'
         '401':
@@ -26,147 +33,11 @@ paths:
           description: 'Not found'
         '500':
           description: 'Internal server error'
-    post:
-      tags:
-        - items
-      summary: Create an item
-      operationId: createItem
-      requestBody:
-        description: create an item
-        content:
-          application/json:
-            schema:
-              $ref: '#/components/schemas/ItemRequestBody'
-      responses:
-        '201':
-          description: 'Created'
-        '400':
-          description: 'Bad request'
-        '401':
-          description: 'Unauthorized'
-        '403':
-          description: 'Forbidden'
-        '404':
-          description: 'Not found'
-        '500':
-          description: 'Internal server error'
-
-  /api/v1/items/{id}:
-    get:
-      summary: Get item by id
-      operationId: getItemById
-      tags:
-        - items
-      parameters:
-        - name: id
-          in: path
-          required: true
-          description: The id of the item to retrieve
-          schema:
-            type: integer
-            format: int64
-      responses:
-        '200':
-          description: Success
-          content:
-            application/json:
-              schema:
-                $ref: "#/components/schemas/Item"
-        '400':
-          description: 'Bad request'
-        '401':
-          description: 'Unauthorized'
-        '403':
-          description: 'Forbidden'
-        '404':
-          description: 'Not found'
-        '500':
-          description: 'Internal server error'
-    put:
-      tags:
-        - items
-      summary: Update an item
-      operationId: updateItem
-      parameters:
-        - name: id
-          in: path
-          required: true
-          description: The id of the item to be updated
-          schema:
-            type: integer
-            format: int64
-      requestBody:
-        description: update an existing item
-        content:
-          application/json:
-            schema:
-              $ref: '#/components/schemas/ItemRequestBody'
-      responses:
-        '204':
-          description: No content
-        '400':
-          description: 'Bad request'
-        '401':
-          description: 'Unauthorized'
-        '403':
-          description: 'Forbidden'
-        '404':
-          description: 'Not found'
-        '500':
-          description: 'Internal server error'
-    delete:
-      summary: Delete item by id
-      operationId: deleteItem
-      tags:
-        - items
-      parameters:
-        - name: id
-          in: path
-          required: true
-          description: The id of the item to be deleted
-          schema:
-            type: integer
-            format: int64
-      responses:
-        '204':
-          description: No content
-        '400':
-          description: 'Bad request'
-        '401':
-          description: 'Unauthorized'
-        '403':
-          description: 'Forbidden'
-        '404':
-          description: 'Not found'
-        '500':
-          description: 'Internal server error'
-
 components:
   schemas:
-    Item:
+    CategoryOfLaw:
       required:
-        - id
-        - name
-        - description
+        - categoryOfLawCode
       properties:
-        id:
-          type: integer
-          format: int64
-        name:
+        categoryOfLawCode:
           type: string
-        description:
-          type: string
-    ItemRequestBody:
-      required:
-        - name
-        - description
-      properties:
-        name:
-          type: string
-        description:
-          type: string
-
-    Items:
-      type: array
-      items:
-        $ref: "#/components/schemas/Item"
