+++ scheme-service/src/main/java/uk/gov/justice/laa/fee/scheme/controller/FeeCalculationController.java
@@ -1,4 +1,4 @@
-package uk.gov.justice.laa.fee.scheme.feecalculator.controller;
+package uk.gov.justice.laa.fee.scheme.controller;
 
 import lombok.AllArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -6,9 +6,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import uk.gov.justice.laa.fee.scheme.feecalculator.model.FeeCalculationRequestDto;
-import uk.gov.justice.laa.fee.scheme.feecalculator.model.FeeCalculationResponseDto;
-import uk.gov.justice.laa.fee.scheme.feecalculator.service.FeeService;
+import uk.gov.justice.laa.fee.scheme.model.FeeCalculationRequestDto;
+import uk.gov.justice.laa.fee.scheme.model.FeeCalculationResponseDto;
+import uk.gov.justice.laa.fee.scheme.service.FeeService;
 
 @RestController
 @AllArgsConstructor
