+++ scheme-service/src/main/java/uk/gov/justice/laa/fee/scheme/model/FeeCalculation.java
@@ -1,4 +1,4 @@
-package uk.gov.justice.laa.fee.scheme.feecalculator.model;
+package uk.gov.justice.laa.fee.scheme.model;
 
 import lombok.Builder;
 import lombok.Data;
